requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
certifi==2024.8.30
click==8.1.7
deprecation==2.1.0
fastapi==0.95.2
frozenlist==1.4.1
gotrue==2.9.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
joblib==1.4.2
mangum==0.17.0
multidict==6.1.0
numpy==1.21.6
packaging==24.1
pandas==1.3.5
postgrest==0.17.1
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
realtime==2.0.5
scikit-learn==1.0.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.8.1
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.6.1
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.31.1
websockets==13.1
yarl==1.14.0
